move track structs around
